Move scorpio interface outside of mct_coupling subdirectory to its own directory under src/share/
