+++ components/scream/src/share/CMakeLists.txt
@@ -26,3 +26,4 @@ set_target_properties(scream_share PROPERTIES
 if (NOT SCREAM_LIB_ONLY)
   add_subdirectory(tests)
 endif()
+  add_subdirectory(io)
