+++ components/scream/src/share/io/tests/scorpio_tests.cpp
@@ -2,7 +2,7 @@
 
 #include "scream_config.h"
 #include "ekat/ekat_pack.hpp"
-#include "mct_coupling/scream_scorpio_interface.hpp"
+#include "share/io/scream_scorpio_interface.hpp"
 #include "ekat/mpi/ekat_comm.hpp"
 #include "share/scream_types.hpp"
 #include "ekat/util/ekat_md_array.hpp"
